+++ github-crawler-core/src/test/java/com/societegenerale/githubcrawler/output/CIdroidReadyCsvFileOutputTest.kt
@@ -13,26 +13,30 @@ class CIdroidReadyCsvFileOutputTest {
     @Test
     fun shouldGenerateExpectedFile() {
 
-        val cIdroidReadyCsvFileOutput = CIdroidReadyCsvFileOutput(listOf("indic1", "indic2"))
+        val cIdroidReadyCsvFileOutput = CIdroidReadyCsvFileOutput(listOf("indic1", "indic2","task1","task2"))
 
         val repo1masterBranchIndicators = mapOf(Pair("indic1", "value1_forRepo1"), Pair("indic2", "value2_forRepo1"))
         val repo1Indicators = mapOf(Pair(Branch("master"), repo1masterBranchIndicators))
-        val repo1 = buildRepoWithIndicators("orga/repo1", repo1Indicators)
+        val repo1 = buildRepoWithIndicators("orga/repo1", repo1Indicators, emptyMap())
 
-        val repo2 = buildRepoWithIndicators("orga/repo2", emptyMap())
+        val repo2masterBranchtasksIndicator = mapOf(Pair("task1", "value1_forRepo2"), Pair("task2", "value2_forRepo2"))
+        val repo2TaskResults= mapOf(Pair(Branch("master"), repo2masterBranchtasksIndicator))
+
+        val repo2 = buildRepoWithIndicators("orga/repo2", emptyMap(),repo2TaskResults)
 
         val repo3someBranchIndicators = mapOf(Pair("indic1", "value1_forRepo3"), Pair("indic2", "value2_forRepo3"), Pair("indic3", "value3_forRepo3"))
         val repo3Indicators = mapOf(Pair(Branch("someBranch"), repo3someBranchIndicators))
-        val repo3 = buildRepoWithIndicators("orga/repo3", repo3Indicators)
+        val repo3 = buildRepoWithIndicators("orga/repo3", repo3Indicators, emptyMap())
 
 
         cIdroidReadyCsvFileOutput.output(repo1);
         cIdroidReadyCsvFileOutput.output(repo2);
         cIdroidReadyCsvFileOutput.output(repo3);
 
-        val expectedFileContent = "repositoryFullName;branchName;indic1;indic2" + System.lineSeparator() +
-                "orga/repo1;master;value1_forRepo1;value2_forRepo1;" + System.lineSeparator() +
-                "orga/repo3;someBranch;value1_forRepo3;value2_forRepo3;" + System.lineSeparator()
+        val expectedFileContent = "repositoryFullName;branchName;indic1;indic2;task1;task2" + System.lineSeparator() +
+                "orga/repo1;master;value1_forRepo1;value2_forRepo1;N/A;N/A;" + System.lineSeparator() +
+                "orga/repo2;master;N/A;N/A;value1_forRepo2;value2_forRepo2;" + System.lineSeparator()+
+                "orga/repo3;someBranch;value1_forRepo3;value2_forRepo3;N/A;N/A;" + System.lineSeparator()
 
         val generatedFile = File(".").walkBottomUp()
                 .filter { f -> f.isFile }
@@ -42,7 +46,7 @@ class CIdroidReadyCsvFileOutputTest {
 
     }
 
-    private fun buildRepoWithIndicators(fullName: String, indicators: Map<Branch, Map<String, String>>): Repository {
+    private fun buildRepoWithIndicators(fullName: String, indicators: Map<Branch, Map<String, String>>, repoTaskResults: Map<Branch, Map<String, String>>): Repository {
 
         return Repository(name = "dummyRepo",
                 creationDate = Date(),
@@ -53,7 +57,8 @@ class CIdroidReadyCsvFileOutputTest {
                 reason = null,
                 url = "http://hello",
                 indicators = indicators,
-                branchesToParse = listOf(Branch("master"))
+                miscTasksResults = repoTaskResults,
+                branchesToParse = setOf(Branch("master"))
         )
 
     }